wx.subscription.web 模块使用统一测试入口
