+++ wx-subscription-web/src/test/resources/log4j.properties
@@ -1,0 +1,42 @@
+#设置级别和目的地  
+log4j.rootLogger=DEBUG, File, Console
+#输出到控制台
+log4j.appender.Console=org.apache.log4j.ConsoleAppender
+#设置输出样式
+log4j.appender.Console.layout=org.apache.log4j.PatternLayout
+#自定义样式
+# %r 时间 0
+# %t 方法名 main
+# %p 优先级 DEBUG/INFO/ERROR
+# %c 所属类的全名(包括包名)
+# %l 发生的位置，在某个类的某行
+# %m 输出代码中指定的讯息，如log(message)中的message
+# %n 输出一个换行符号
+#log4j.appender.Console.layout.ConversionPattern=[%p][%d{yyyy-MM-dd HH:mm:ss:SSS}][%C-%M] %m%n
+log4j.appender.Console.layout.ConversionPattern=[%p][%d{MM-dd HH:mm:ss}][%c{1}] %m%n
+log4j.logger.org.apache.commons=INFO
+log4j.logger.org.springframework=INFO
+log4j.logger.org.hibernate=WARN
+log4j.logger.org.hibernate.SQL=INFO
+log4j.logger.org.hibernate.hql=WARN
+log4j.logger.org.hibernate.jdbc=WARN
+log4j.logger.org.hibernate.cfg=ERROR
+log4j.logger.org.mybatis.spring=WARN
+log4j.logger.com.zaxxer.hikari.pool=INFO
+log4j.logger.com.zaxxer.hikari=INFO
+log4j.logger.org=ERROR
+#输出到文件(这里默认为追加方式)
+#log4j.appender.File=org.apache.log4j.DailyRollingFileAppender
+log4j.appender.File=org.apache.log4j.RollingFileAppender
+log4j.appender.File.MaxBackupIndex=10
+log4j.appender.File.MaxFileSize=20480KB
+#log4j.appender.File.Threshold=debug
+#log4j.appender.File.Append=true
+#设置文件输出路径
+log4j.appender.File.File=logs/wx-subscription.log
+#设置文件输出样式
+log4j.appender.File.layout=org.apache.log4j.PatternLayout
+#log4j.appender.File.layout.ConversionPattern=[%p][%d{yyyy-MM-dd HH:mm:ss:SSS}][%C-%M] %m%n
+log4j.appender.File.layout.ConversionPattern=[%p][%d{yyyy-MM-dd HH:mm:ss:SSS}][%c{1}-%M] %m%n
+#设置编码字符集
+log4j.appender.File.encoding=UTF-8
